requirements.txt
pyserial~=3.4
pyautogui~=0.9.53
marshmallow~=3.7.1
marshmallow_dataclass
marshmallow_enum
